Move image loading routines into new class
Extract the LoadMemDump and LoadLdPic (and associated bit reading
routines) from the main Beebview class and put them in a new
BbcImageLoader class.  Also add a LoadAuto function to attempt loading
with LdPic but then fall back to LoadMemDump.

Also take the opportunity to pass a pointer to a byte array to the image
loader (rather than a stream), as this allows us to just memory map the
input file rather than reading it a byte at a time.