Add support for loading compressed ScrLoad images
Add code to load images compressed with ScrSave/ScrLoad run-length
compression.  This doesn't handle setting the mode and palette as this
is stored in the load and execute addresses which aren't available away
from the BBC.

Uncompressed ScrSave/ScrLoad images are just memory dumps, so no
additional code is required to handle them.