Move author copyright to individual files
Instead of trying to handle the list of authors separately to the file
copyright, remove AUTHORS.md and change the copyright in all the files
to the contributor names.  This helps to keep the author copyright with
the file content if any of the files are used elsewhere.

While updating the copyright headers, also correct the copyright dates
as a number of these were incorrect and add a missing copyright header
to the CI workflow file.

Also take the opportunity to reduce the size of the file headers by
removing the GPL boilerplate from each file and replacing it with a
SPDX-License-Identifier header line.