Enable and add a range of Overcommit hooks
Enable some out of the box Overcommit rules for checking trailing
whitespace and running yamllint.  Add three custom hooks which ensure
files end with a single newline, are encoded as UTF-8 without a BoM and
have up-to-date copyright notices.

Fix one issue that the hooks picked up with a missing document start in
the CI workflow file.